+++ src/main.py
@@ -10,32 +10,6 @@ from download_data import *
 from functions_background import *
 
 
-class SlyProgress:
-    def __init__(self, api, task_id, pbar_element_name):
-        self.api = api
-        self.task_id = task_id
-        self.pbar_element_name = pbar_element_name
-        self.pbar = None
-        # self.refresh_params('-', 0)
-
-    def refresh_params(self, desc, total):
-        self.pbar = sly.Progress(desc, total)
-
-    def refresh_progress(self):
-        fields = [
-            {"field": f"data.{self.pbar_element_name}", "payload": int(self.pbar.current * 100 / self.pbar.total)},
-            {"field": f"data.{self.pbar_element_name}Message", "payload": self.pbar.message},
-            {"field": f"data.{self.pbar_element_name}Current", "payload": self.pbar.current},
-            {"field": f"data.{self.pbar_element_name}Total", "payload": self.pbar.total},
-        ]
-        self.api.task.set_fields(self.task_id, fields)
-
-    def next_step(self):
-        self.pbar.iter_done_report()
-        self.refresh_progress()
-
-
-
 
 @app.callback("preview")
 @sly.timeit
@@ -50,14 +24,14 @@ def preview(api: sly.Api, task_id, context, state, app_logger):
     api.task.set_fields(task_id, fields)
 
     process_video(sly_progress, state, is_preview=True)
+    sly_progress.refresh_params('Uploading video', 1)
 
     video_path = os.path.join(app.data_dir, './preview.mp4')
     remote_video_path = os.path.join(f"/SLYvSynth/{task_id}", "preview.mp4")
     if api.file.exists(team_id, remote_video_path):
         api.file.remove(team_id, remote_video_path)
 
-    sly_progress.refresh_params('Uploading video', 1)
-
+    # file_info = api.file.upload(team_id, video_path, remote_video_path, progress_cb=sly_progress.pbar.iters_done_report)
     file_info = api.file.upload(team_id, video_path, remote_video_path)
 
     sly_progress.next_step()
@@ -89,9 +63,8 @@ def synthesize(api: sly.Api, task_id, context, state, app_logger):
     res_project = api.project.get_info_by_id(res_project_id)  # load full project info
 
     fields = [
-
-        {"field": "data.step4Loading", "payload": False},
-        {"field": "data.done4", "payload": True},
+        {"field": "data.step5Loading", "payload": False},
+        {"field": "data.done5", "payload": True},
 
 
         {"field": "data.dstProjectId", "payload": res_project.id},
