+++ src/ui/augmentations.html
@@ -1,9 +1,9 @@
-<sly-card title="Augmentations" class="mb15"
+<sly-card title="Augmentations"
 
           subtitle="Choose one of the prepared templates or provide custom pipeline"
-          :disabled="{disabled: state.disabled2, message: 'complete stage 1 to unlock'}"
           :options="{collapsable: true}"
-          :collapsed.sync="state.collapsed2"
+          :collapsed.sync="state.collapsed3"
+          :disabled="{disabled: state.disabled3, message: 'complete stage 2 to unlock'}"
           ref="step1" style="width: 100%"
 >
 
@@ -296,26 +296,20 @@
             </el-tab-pane>
         </el-tabs>
     </div>
-
-</sly-card>
-
-
-<div v-if="data.done1" class="fflex m15 mb15" style="width: 100%;">
-    <sly-card style="width: 100%;">
-        <div>
-            <div class="fflex mt10" v-if="!data.done2">
-                <el-button type="primary" class="regular" @click="command('apply_augs')">Apply augmentations
-                </el-button>
-            </div>
-            <div v-if="data.done2">
-                <i class="zmdi zmdi-check-circle mr5 mt10" style="color: #13ce66"></i>
-                <span style="color: #5a6772;">Augmentations has been successfully defined</span>
-            </div>
-            <el-button type="warning" class="regular mt10" :plain="true"
-                       @click="state.restartFrom = 2;"
-                       v-if="data.done2">
-                <i class="zmdi zmdi-rotate-left mr5"></i> reselect
+    <div>
+        <div class="fflex mt10" v-if="!data.done3">
+            <el-button type="primary" class="regular" @click="command('apply_augs')">Apply augmentations
             </el-button>
         </div>
-    </sly-card>
-</div>
+        <div v-if="data.done3">
+            <i class="zmdi zmdi-check-circle mr5 mt10 mb10" style="color: #13ce66"></i>
+            <span style="color: #5a6772;">Augmentations has been successfully defined</span>
+        </div>
+        <el-button type="warning" class="regular mt10" :plain="true"
+                   @click="state.restartFrom=3;"
+                   v-if="data.done3">
+            <i class="zmdi zmdi-rotate-left mr5"></i> reselect
+        </el-button>
+    </div>
+</sly-card>
+
