+++ src/ui/input_backgrounds.html
@@ -1,0 +1,66 @@
+<sly-card
+        title="Select backgrounds"
+
+        subtitle="Download backgrounds from server to local app directory"
+        :options="{collapsable: true}"
+        :collapsed.sync="state.collapsed1"
+        ref="step1"
+>
+
+
+    <sly-select-team-workspace
+            :team-id.sync="state.bgTeamId"
+            :workspace-id.sync="state.bgWorkspaceId"
+            :disabled="data.done1">
+
+    </sly-select-team-workspace>
+    <sly-select-project
+            :workspace-id="state.bgWorkspaceId"
+            :project-id.sync="state.bgProjectId"
+            :disabled="data.done1">
+    </sly-select-project>
+
+    <el-checkbox class="mt15" v-model="state.allDatasets"
+                 :disabled="data.done1">All datasets
+    </el-checkbox>
+    <sly-select-dataset
+            v-show="!state.allDatasets"
+            :project-id="state.bgProjectId"
+            :datasets.sync="state.bgDatasets"
+            :options="{'multiple': true}"
+            :disabled="data.done1">
+    </sly-select-dataset>
+
+    <div class="mt15">
+        <el-button
+                type="primary"
+                class="regular"
+                :disabled="!state.bgProjectId || (!state.allDatasets && !(state.bgDatasets.length > 0))"
+                :loading="data.step1Loading"
+                @click="data.step1Loading=true; command('download_backgrounds')"
+                v-if="!data.done1">
+            Download
+
+
+        </el-button>
+        <div v-if="data.done1">
+            <i class="zmdi zmdi-check-circle mr5 mt10 mb10" style="color: #13ce66"></i>
+            <span style="color: #5a6772;">Backgrounds has been successfully downloaded</span>
+        </div>
+        <div v-if="data.step1Loading" class="mt15">
+            <div>{{ data.progressDownloadBackgroundsMessage }}: <b><span
+                    style="color: #20a0ff">{{ data.progressDownloadBackgroundsCurrent }}</span> /
+                {{ data.progressDownloadBackgroundsTotal }}</b></div>
+            <el-progress
+                    :percentage="Math.round(data.progressDownloadBackgroundsCurrent * 100
+                     / data.progressDownloadBackgroundsTotal)"></el-progress>
+        </div>
+
+        <el-button type="warning" class="regular mt10" :plain="true"
+                   @click="state.restartFrom = 1;"
+                   v-if="data.done1">
+            <i class="zmdi zmdi-rotate-left mr5"></i> reselect
+        </el-button>
+    </div>
+
+</sly-card>
